Correção do tamnaho da tela de cadastro de produtos

+++ app/models/ficha_tecnica.py
@@ -15,7 +15,7 @@ class FichaTecnica(Base):
     id = Column(Integer, primary_key=True, autoincrement=True)
     produto_id = Column(Integer, ForeignKey("produtos.id"), nullable=False)
     codigo_produto = Column(String(50), nullable=False)  # ← ADICIONADO
-    sacos_batida = Column(Integer, nullable=False, default=0)
+    sacos_batida = Column(Numeric(12, 4), nullable=False, default=0)
 
     itens = relationship(
         "ItemFichaTecnica",
